Separate class switching action into individual module
